requirements.txt
biopython==1.79
numpy==1.21.1
pandas==1.3.1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
scipy==1.7.1